+++ QSDFT_FMT_Slit_alpha_v.0.1.4_float/run.sh
@@ -4,10 +4,9 @@
 # 1.18 1.35 1.61 1.93 2.31 2.90 3.63 4.75 6.51 10.0
 #)
 
-if [ ! -e qsdft_fmt_slit.exe ]; then
-	#c++ -O2 qsdft_fmt_slit.cpp -o qsdft_fmt_slit.exe
-	c++ -O2 qsdft_fmt_slit_steele.cpp -o qsdft_fmt_slit.exe
-fi
+#if [ ! -e qsdft_fmt_slit_steele.exe ]; then
+	c++ -O2 qsdft_fmt_slit_steele.cpp -o qsdft_fmt_slit_steele.exe
+#fi
 
 if [ ! -d results ]; then
 	mkdir results
@@ -22,9 +21,19 @@ for ((i=42;i<=1000;i+=2)); do
 	H=`awk -v w=${w} "BEGIN {print w+0.195}"`
 	echo "Slit width = ${H} [nm]"
 	sed -i "s/XXX/${H}/g" parameters.txt
-	./qsdft_fmt_slit.exe
-	mv PP0_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
-	mv PP0_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	./qsdft_fmt_slit_steele.exe
+	if [ -e PP0_vs_Vgamma_data_vs.txt ]; then
+	  mv PP0_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv PP0_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
+	if [ -e Pa_vs_Vgamma_data_vs.txt ]; then
+	  mv Pa_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv Pa_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
+	if [ -e atm_vs_Vgamma_data_vs.txt ]; then
+	  mv atm_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv atm_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
 	cp ${w}_data_vs.txt ./results/${w}_data_vs.txt
 	cp ${w}_data_ls.txt ./results/${w}_data_ls.txt
 	rm parameters.txt
